+++ onecloud/add-node.yml
@@ -5,5 +5,5 @@
 - hosts: worker_nodes
   roles:
     - worker-node
-    - utils/gpu-init
+    - { role: utils/gpu-init, when: "join_as_host|default(false)" }
     - utils/kernel-check
