+++ onecloud/roles/utils/kernel-check/tasks/main.yml
@@ -100,7 +100,7 @@
   when:
   - is_yunion_kernel_running.rc != 0
   - ansible_connection == "ssh"
-  - k8s_node_as_oc_host|default(false)|bool == true
+  - k8s_node_as_oc_host|default(false)|bool == true or enable_lbagent|default(false)|bool == true
   - is_centos_based is defined
   - is_kylin_based|default(false)|bool == false
