+++ onecloud/roles/worker-node/lbagent/templates/lbagent.conf.j2
@@ -1,0 +1,2 @@
+listen_interface = "{{ lbagent_listen_interface }}"
+access_ip = "{{ lbagent_access_ip }}"
